Move setup_parser to mask/config.py
