+++ src/bentopy/mask/config.py
@@ -1,0 +1,86 @@
+import argparse
+from pathlib import Path
+
+DEFAULT_CONTAINMENT_RESOLUTION = 1.0
+DEFAULT_MASK_RESOLUTION = 0.5
+
+
+def process_labels(labels: str) -> list[int]:
+    return [int(label) for label in labels.split(",")]
+
+
+def setup_parser(parser=None):
+    if parser is None:
+        parser = argparse.ArgumentParser(
+            description="Set up masks based on structures and compartment segmentations.",
+            prog="mask",
+            epilog='"Relabel, relabel, relabel."',
+        )
+    parser.add_argument(
+        "input",
+        type=Path,
+        help="Input structure file to subject to segmentation.",
+    )
+    parser.add_argument(
+        "output",
+        type=Path,
+        help="Output path for the resulting voxel mask.",
+    )
+    parser.add_argument(
+        "--no-interactive",
+        dest="interactive",
+        action="store_false",
+        help="""Do not ask for command line input at runtime. May be desirable 
+        when all inputs are known and in a scripted context.
+        (normally interactive)""",
+    )
+    parser.add_argument(
+        "--containment-resolution",
+        default=DEFAULT_CONTAINMENT_RESOLUTION,
+        type=float,
+        help="Resolution for the compartment finding routine (nm). (default: %(default)s nm)",
+    )
+    parser.add_argument(
+        "--mask-resolution",
+        default=DEFAULT_MASK_RESOLUTION,
+        type=float,
+        help="Voxel size (resolution) for the exported mask (nm). (default: %(default)s nm)",
+    )
+    parser.add_argument(
+        "--selection",
+        default="not resname W ION",
+        type=str,
+        help="MDAnalysis selection string for the atom group over which to perform the segmentation. (default: '%(default)s')",
+    )
+    parser.add_argument(
+        "--labels",
+        type=process_labels,
+        help="Pre-selected compartment labels, comma-separated.",
+    )
+    parser.add_argument(
+        "--plot",
+        type=Path,
+        nargs="?",
+        const="containment.html",
+        help="""Optional output path for the visualization page generated by 
+        mdvcontainment during segmentation. 
+        (when used, default: %(const)s).""",
+    )
+    parser.add_argument(
+        "-b",
+        "--inspect-labels-path",
+        type=Path,
+        nargs="?",
+        const="labels.gro",
+        help="""Optional output path to a gro file to write labeled voxel positions to.
+        This file can be inspected with molecule viewers, which is very helpful 
+        in determining which labels match the compartments you want to select.
+        (when used, default: %(const)s)""",
+    )
+    parser.add_argument(
+        "--debug-voxels",
+        type=Path,
+        help="""Write the final voxel mask as a gro file for inspection with molecule viewers. 
+        This can be useful when you want to verify the voxel mask that is produced for some selection of labels.""",
+    )
+    return parser
